split algorithms, changed list visualizer and fixed issue with list generation
