+++ src/components/ListConfig.jsx
@@ -11,21 +11,26 @@ class ListConfig extends Component {
   render() {
     return (
         <Card className="appConfig white">
-          <label>
-            List Elements:
-            <input type="number" id="height" name="height" min="0" max="100" value={this.props.size} placeholder="10" onChange={this.props.handleSizeChange} />
-          </label>
-          <div className="btnContainer">
-            <Button className="blue" waves='light' onClick={this.props.updateListSize}>
-              Load
-            </Button>
-            <Button className="blue" waves='light' onClick={this.props.listRandomizer}>
-              Randomize
-            </Button>
-          </div>
-          <div className="listValues">
-            [{ this.props.lst[0].map((elem) => <span className="listValuesItem" key={shortid.generate()}>{elem}</span>)}]
-          </div>        
+          <section className="content">
+            <label>
+              List Elements:
+              <input type="number" id="height" name="height" min="0" value={this.props.size} placeholder="10" onChange={this.props.handleSizeChange} />
+            </label>
+            <div className="btnContainer">
+              <Button className="blue" waves='light' onClick={this.props.updateListSize}>
+                Load
+              </Button>
+              <Button className="blue" waves='light' onClick={this.props.listRandomizer}>
+                Randomize
+              </Button>
+            </div>
+            <section className="list-view">
+              
+              <div className="listValues">
+                [{ this.props.lst[0].map((elem) => <span className="listValuesItem" key={shortid.generate()}>{elem}</span>)}]
+              </div>  
+            </section>
+          </section>
         </Card>
     );
   }
