+++ src/components/ListConfig.scss
@@ -5,21 +5,30 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
-    align-items: center;
+    align-items: center;  
+}
+.content {
+    max-width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;  
+}
+
+.list-view {
+    width: 100%;
+    max-height: 30vh;
+    overflow-y: auto;
 }
 
 .listValues {
-    width: 100%;
     display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    text-align: center;
-    overflow: hidden;
-    overflow-x: auto;   
+    flex-wrap: wrap;
+    justify-content: center;
     margin: 5px;
     padding: 5px;
     padding-left: 5px;
+    
 }
 
 .listValuesItem {
@@ -36,6 +45,7 @@
 
 .btnContainer {
     width: 100%;
+    margin-bottom: 5vh;
     display: flex;
     flex-direction: row;
     justify-content: center;
