+++ src/components/ListElement.css
@@ -1,7 +1,7 @@
 .list-element {
-    width: 2.5vw;
-    margin-left: 5px;
-    margin-right: 5px;
+    width: 3px;
+    margin-left: 2.5px;
+    margin-right: 2.5px;
     background-color: black;
     transition-duration: 0.1s;
 }
